src/requirement.txt
absl-py==0.7.0
astor==0.7.1
autopep8==1.4.3
certifi==2019.3.9
cycler==0.10.0
EasyProcess==0.2.5
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
numpy==1.16.2
olefile==0.46
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pbr==5.1.3
Pillow==5.4.1
protobuf==3.6.1
pycodestyle==2.5.0
pynput==1.4.2
pyparsing==2.4.0
pyscreenshot==0.5.1
python-dateutil==2.8.0
python-xlib==0.25
pytz==2018.9
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
Werkzeug==0.14.1
